requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
install==1.3.4
Pillow==8.2.0
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
soupsieve==2.2.1
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.4
webencodings==0.5.1
